refactor(rx-store): reorganize files
