+++ libs/rx-store/src/index.ts
@@ -1,7 +1,7 @@
 export * from './models/change';
-export * from './stores/rx-arrays';
-export * from './stores/rx-connect-arrays';
-export * from './stores/rx-connect-map';
-export * from './stores/rx-map';
-export * from './stores/rx-set';
-export * from './stores/validity-map';
+export * from './structures/rx-arrays';
+export * from './structures/rx-connect-arrays';
+export * from './structures/rx-connect-map';
+export * from './structures/rx-map';
+export * from './structures/rx-set';
+export * from './utils/validity-map';
