+++ libs/rx-store/src/utils/map-keys-to-values.ts
@@ -1,7 +1,7 @@
 import { isNotUndefined } from '@ns3/ts-utils';
 import { combineLatest, Observable, of } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
-import { RxMap } from '../stores/rx-map';
+import { RxMap } from '../structures/rx-map';
 
 export function mapKeysToValues<TKey, TValue>(itemsMap: RxMap<TKey, TValue>) {
   return (
